fix: moved supported chains to constants folder
