+++ src/constants/chains.ts
@@ -1,0 +1,7 @@
+
+export enum SupportedChainId {
+    CELO_MAINNET = 42220,
+    GOERLI_TESTNET = 5,
+    OPTIMISM_MAINNET = 10,
+    POLYGON_MAINNET = 137,
+}
