+++ src/types/zeroWallet.ts
@@ -1,11 +1,5 @@
 import { ethers } from 'ethers';
-
-export enum SupportedChainId {
-  CELO_MAINNET = 42220,
-  GOERLI_TESTNET = 5,
-  OPTIMISM_MAINNET = 10,
-  POLYGON_MAINNET = 137,
-}
+import { SupportedChainId } from '../constants/chains';
 
 export type ZeroWalletProviders = {
   [key in SupportedChainId]?: ethers.providers.Provider;
